app/requirements.txt
cryptography==44.0.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
tk==0.1.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
